fix(clippy): resolve all remaining lint warnings across workspace
- Add `#![allow(dead_code)]` to daemon and gui crates for unimplemented stubs
- Derive `Default` for `FolderScheduleState` instead of manual impl
- Use `std::io::Error::other()` shorthand over `Error::new(Other, ...)`
- Add `.truncate(false)` to explicit file opens to satisfy clippy::suspicious_open_options
- Elide redundant explicit lifetimes in gui view signatures
- Replace `PADDING as u16 * 2` with `PADDING * 2` (const already u16)
- Use `checked_div` for progress percentage to avoid potential div-by-zero lint
- Use `dir.keep()` instead of deprecated `into_path()` in vfs_factory test helper
- Suppress unused `Task` return values in gui update tests with `let _ =`
- Use `assert!()` instead of `assert_eq!(x, true)` in dbus tests
- Gate `use camino::Utf8PathBuf` behind `#[cfg(target_os = "macos")]`
- Add `#[allow(dead_code)]` to non-Windows placeholder function

+++ crates/gui/src/views/add_account.rs
@@ -46,6 +46,6 @@ pub fn add_account_view<'a>(url_input: &'a str, error: Option<&'a str>) -> Eleme
         .height(Length::Fill)
         .center_x(Length::Fill)
         .center_y(Length::Fill)
-        .padding(PADDING as u16 * 2)
+        .padding(PADDING * 2)
         .into()
 }
